Prototipo App/requirements.txt
# requirements.txt

streamlit
pydicom
opencv-python
Pillow
scikit-image
albumentations
numpy
scipy
matplotlib
pandas
joblib
tqdm
streamlit-option-menu
streamlit-lottie
watchdog
streamlit_toggle
transformers
torch
torchvision
torchaudio
seaborn
fpdf
